+++ CHANGELOG.md
@@ -6,6 +6,11 @@ ByteriX, 2017. All right reserved.
 
 # Versions
 
+## 2.1.1 (23.03.2017)
+##### Improvements
+renamed globalCellHeight to cellHeight in BxInputSettings
+added headerHeight & footerHeight to BxInputSettings as global values of content size
+
 ## 2.1.0 (23.03.2017)
 ##### Change logic
 * extracted BxInputValueRow for didChangedValue